requirements.txt
selene>=2.0.0rc9
pytest~=8.2.2
allure-pytest==2.13.5
python-dotenv==1.0.1
selenium==4.23.1
openpyxl==3.1.5
webdriver-manager==4.0.2
python-dotenv==1.0.1
